+++ src/js/MyNotes.js
@@ -2,24 +2,19 @@ import $ from 'jquery';
 
 class MyNotes {
     constructor() {
-        this.editButton = document.querySelectorAll('.edit-note');
-        this.deleteButton = document.querySelectorAll('.delete-note');
-        this.updateButton = document.querySelectorAll('.update-note');
+        this.myNotes = document.querySelector('#my-notes');
         this.events();
     }
 
     events = () => {
-        this.editButton.forEach(btn => btn.addEventListener('click', (e) => {
-            this.editNote(e);
-        }));
+        this.myNotes.addEventListener("click", e => this.clickHandler(e));
+        document.querySelector('.submit-note').addEventListener('click', (e) => this.createNote(e));
+    }
 
-        this.deleteButton.forEach(btn => btn.addEventListener('click', (e) => {
-            this.deleteNote(e);
-        }));
-
-        this.updateButton.forEach(btn => btn.addEventListener('click', (e) => {
-            this.updateNote(e);
-        }));
+    clickHandler(e) {
+        if (e.target.classList.contains("delete-note") || e.target.classList.contains("fa-trash-o")) this.deleteNote(e);
+        if (e.target.classList.contains("edit-note") || e.target.classList.contains("fa-pencil") || e.target.classList.contains("fa-times")) this.editNote(e);
+        if (e.target.classList.contains("update-note") || e.target.classList.contains("fa-arrow-right")) this.updateNote(e);
     }
 
     findNearestParentLi(el) {
@@ -56,7 +51,6 @@ class MyNotes {
         thisNote.querySelector('.note-body-field').classList.remove('note-active-field');
 
         thisNote.querySelector('.update-note').classList.remove('update-note--visible');
-
     }
 
     editNote = (e) => {
@@ -72,7 +66,7 @@ class MyNotes {
     deleteNote = async (e) => {
         const thisNote = this.findNearestParentLi(e.target);
 
-        const result = await request('DELETE', null, thisNote)
+        const result = await requestToDB('DELETE', null, thisNote)
         console.log(result);
 
         $(thisNote).slideUp();
@@ -85,18 +79,51 @@ class MyNotes {
             title: thisNote.querySelector('.note-title-field').value,
             content: thisNote.querySelector('.note-body-field').value
         }
-        const result = await request('PUT', content, thisNote);
+        const result = await requestToDB('PUT', content, thisNote);
         console.log(result);
 
         this.saveMode(thisNote);
+    }
+
+    createNote = async (e) => {
+        const content = {
+            title: document.querySelector('.new-note-title').value,
+            content: document.querySelector('.new-note-body').value,
+            status: 'private'
+        }
+        if (content.title && content.content) {
+            const result = await requestToDB('POST', content, null);
+            console.log(result);
+
+            document.querySelector('.new-note-title').value = '';
+            document.querySelector('.new-note-body').value = '';
+
+            document.querySelector('#my-notes').insertAdjacentHTML('afterBegin', `
+                <li data-note-id="${result.id}">
+                    <input class="note-title-field" value="${result.title.raw}" readonly>
+                    <span class="edit-note">
+                        <i class="fa fa-pencil" aria-hidden="true"></i>Edit
+                    </span>
+                    <span class="delete-note">
+                        <i class="fa fa-trash-o" aria-hidden="true"></i>Delete
+                    </span>
+                    <textarea class="note-body-field" readonly>
+                        ${result.content.raw}
+                    </textarea>
+                    <span class="update-note btn btn--blue btn--small">
+                        <i class="fa fa-arrow-right" aria-hidden="true"></i>Save
+                    </span>
+                </li>
+            `);
+        }
     }
 }
 
 export default MyNotes;
 
-async function request(method, content, thisNote) {
+async function requestToDB(method, content, thisNote) {
     const resp = await fetch(
-        `${univerData.root_url}/wp-json/wp/v2/note/${thisNote.getAttribute('data-note-id')}`,
+        `${univerData.root_url}/wp-json/wp/v2/note/${thisNote ? thisNote.getAttribute('data-note-id') : ''}`,
         {
             method,
             headers: {
